Fix overscroll behavior
Previously, I had to disable overscroll, as the "Paper" component didn't
stretch to follow it. I've now reenabled it and made the
background-color of the body match the paper color. It doesn't look
exactly the same, due to some gradients and other shenanigans they're
pulling. Based off of their website, it would appear this is the same
solution they took.