Material design (#1)
* Switch components to material design

It still doesn't look good, but now we have material design in the
project.

* Switch auth page to use more material components

I think I still need to add the theme provider to get margin and padding
automagically, so that's coming up next

* Fix styling issues and fix login page

I finally made it so that the entire page has the material design paper
behind it, which automagically transitions with the user theme.

Additionally, the login page is now using only material components.
There's aditional styling here and there with tailwind utilities
classes, but all focus management, hover transitions, colors, etc is
handled by Material UI. Thank heavens

* Fix overscroll behavior

Previously, I had to disable overscroll, as the "Paper" component didn't
stretch to follow it. I've now reenabled it and made the
background-color of the body match the paper color. It doesn't look
exactly the same, due to some gradients and other shenanigans they're
pulling. Based off of their website, it would appear this is the same
solution they took.